Move engine handler to separate file
